requirements.txt
Pillow==11.0.0
python-dotenv==0.19.0
python-telegram-bot==13.7